+++ app/src/main/java/com/devenlucaz/doscom/MainActivity.kt
@@ -1,11 +1,22 @@
 package com.devenlucaz.doscom
 
-import androidx.appcompat.app.AppCompatActivity
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.provider.Settings
+import androidx.appcompat.app.AppCompatActivity
+import com.devenlucaz.doscom.service.ServiceManager
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        if (Settings.canDrawOverlays(this)) {
+            ServiceManager.startOverlayService(this)
+        } else {
+            val intent = Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION, Uri.parse("package:$packageName"))
+            startActivity(intent)
+        }
     }
 }